Change GraphicsOperations.drawLine(int, int, int, int, color) to (double, double, double, double, color)

+++ uilayer-core/src/main/java/net/avh4/framework/uilayer/scene/GraphicsOperations.java
@@ -14,7 +14,7 @@ public abstract class GraphicsOperations {
 
     public abstract void translate(double deltaX, double deltaY);
 
-    public abstract void drawLine(int startX, int startY, int stopX, int stopY, int argbColor);
+    public abstract void drawLine(double startX, double startY, double stopX, double stopY, int argbColor);
 
     public abstract void drawOval(double leftX, double topY, double width, double height, int argbColor);
